requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
Pillow>=9.0.0

# Augmentation
albumentations>=1.3.0
opencv-python>=4.5.0

# Utilities
tqdm>=4.64.0
matplotlib>=3.6.0
seaborn>=0.11.0

# Optional (for experiment tracking and saving)
tensorboard>=2.10.0

# Safe to pin
scikit-learn>=1.1.0
